服务优化

+++ shop-common/src/main/java/quick/pager/shop/utils/LatitudeLongitudeUtil.java
@@ -43,6 +43,6 @@ public class LatitudeLongitudeUtil {
         double s = 2 * Math.asin(Math.sqrt(Math.pow(Math.sin(a / 2), 2) +
                 Math.cos(radLat1) * Math.cos(radLat2) * Math.pow(Math.sin(b / 2), 2)));
         s = s * EARTH_RADIUS;
-        return Math.round(s * 10000) / 10000;
+        return Math.round(s * 10000) / (double) 10000;
     }
 }
